ImageServiceDocker/requirements.txt
pillow
keras_applications
boto3
